refactor: nextcloud external api client
call nextcloud api first, then create map if everything went well
