+++ backend/src/model/enum/privacy_option.rs
@@ -7,7 +7,7 @@ use utoipa::ToSchema;
 
 /// Enum for map privacy setting options.
 #[typeshare]
-#[derive(Serialize, Deserialize, DbEnum, Debug, ToSchema, Clone)]
+#[derive(Serialize, Deserialize, DbEnum, PartialEq, Debug, ToSchema, Clone)]
 #[ExistingTypePath = "crate::schema::sql_types::PrivacyOption"]
 pub enum PrivacyOption {
     /// Data is private and only visible for owner or members, the data was explicitly shared with.
